+++ src/main/java/bj/highfiveuniversity/book/Services/BookService.java
@@ -2,9 +2,16 @@ package bj.highfiveuniversity.book.Services;
 
 import org.springframework.stereotype.Service;
 
+import bj.highfiveuniversity.book.DTO.BookDTO;
+import bj.highfiveuniversity.book.Exceptions.ResourceNotFoundException;
+import bj.highfiveuniversity.book.Mapper.BookMapper;
+import bj.highfiveuniversity.book.Repository.AuthorRepository;
 import bj.highfiveuniversity.book.Repository.BookRepository;
+import bj.highfiveuniversity.book.models.Author;
 import bj.highfiveuniversity.book.models.Book;
 
+import java.util.ArrayList;
+import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -15,9 +22,13 @@ public class BookService {
     @Autowired
     private final BookRepository bookRepository;
 
+
+    @Autowired
+    private final AuthorRepository authorRepository;
     
-    public BookService(BookRepository bookRepository) {
+    public BookService(BookRepository bookRepository, AuthorRepository authorRepository) {
         this.bookRepository = bookRepository;
+        this.authorRepository = authorRepository;
     }
 
   
@@ -27,11 +38,16 @@ public class BookService {
     }
 
     public void deleteBook(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new ResourceNotFoundException("livre avec l'id " + id + " n'existe pas");
+        }
         bookRepository.deleteById(id);
     }
 
     public void updateBook(Long id, Book book) {
-        Book bookToUpdate = bookRepository.findById(id).get();
+
+        Book bookToUpdate = bookRepository.findById(id)
+        .orElseThrow(() -> new ResourceNotFoundException("livre avec l'id " + id + " n'existe pas"));
         bookToUpdate.setTitle(book.getTitle());
         bookToUpdate.setAuthors(book.getAuthors());
         bookToUpdate.setPublished_at(book.getPublished_at());
@@ -40,16 +56,41 @@ public class BookService {
 
     }
 
-    public Iterable<Book> getBooks() {
-        return bookRepository.findAll();
+    public Iterable<BookDTO> getBooks() {
+
+       Iterable<Book>books = bookRepository.findAll();
+        if (books.spliterator().getExactSizeIfKnown() == 0) {
+            throw new ResourceNotFoundException("Aucun livre retrouvé");
+        }
+        List<BookDTO> bookList = new ArrayList<>();
+
+        for (Book book : books) {
+            bookList.add(BookMapper.toDto(book));
+        }
+
+        return bookList;
     }
     
-    public Book getBookById(Long id) {
-        return bookRepository.findById(id).get();
+    public BookDTO getBookById(Long id) {
+       Book book = bookRepository.findById(id)
+          .orElseThrow(() -> new ResourceNotFoundException("Livre non trouvé avec l'id " + id));
+
+          return BookMapper.toDto(book);
+
     }
 
     public Book getBookByTitle(String title) {
-        return bookRepository.findByTitle(title).get();
+        return bookRepository.findByTitle(title)
+          .orElseThrow(() -> new ResourceNotFoundException("Livre non trouvé avec le titre " + title));
+    }
+
+       public void linkBookToAuthor(Long bookId, Long authorId) {
+        Book book = bookRepository.findById(bookId)
+                .orElseThrow(() -> new ResourceNotFoundException("Livre non trouvé avec l'id " + bookId));
+        Author author = authorRepository.findById(authorId)
+                .orElseThrow(() -> new ResourceNotFoundException("Auteur non trouvé avec l'id " + authorId));
+        book.getAuthors().add(author);
+        bookRepository.save(book);
     }
     
 }
